requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
langchain-text-splitters
langchain-huggingface
sentence-transformers
chromadb
python-dotenv
pypdf
